+++ src/index.css
@@ -1,0 +1,68 @@
+    /* Vendor */
+@import url(../src/vendor/normalize.css);
+@import url(../src/vendor/fonts.css);
+
+    /* Page */
+@import url(../src/blocks/page/page.css);
+@import url(../src/blocks/page/__header/page__header.css);
+@import url(../src/blocks/page/__profile/page__profile.css);
+@import url(../src/blocks/page/__elements/page__elements.css);
+@import url(../src/blocks/page/__footer/page__footer.css);
+
+    /* Popup */
+@import url(../src/blocks/popup/popup.css);
+@import url(../src/blocks/popup/__field/popup__field.css);
+@import url(../src/blocks/popup/__container/popup__container.css);
+@import url(../src/blocks/popup/_opened/popup_opened.css);
+@import url(../src/blocks/popup/__heading/popup__heading.css);
+@import url(../src/blocks/popup/__edit-info/popup__edit-info.css);
+@import url(../src/blocks/popup/__input/popup__input.css);
+@import url(../src/blocks/popup/__submit-button/popup__submit-button.css);
+@import url(../src/blocks/popup/__submit-button/_inactive/popup__submit-button_inactive.css);
+@import url(../src/blocks/popup/__caption/popup__caption.css);
+@import url(../src/blocks/popup/__figure/popup__figure.css);
+@import url(../src/blocks/popup/__image/popup__image.css);
+@import url(../src/blocks/popup/close-button/popup__close-button.css);
+@import url(../src/blocks/popup/__input/_type_error/popup__input_type_error.css);
+@import url(../src/blocks/popup/__input-error/popup__input-error.css);
+@import url(../src/blocks/popup/__input-error/_active/popup__input-error_active.css);
+
+
+@import url(../src/blocks/add-popup/add-popup.css);
+@import url(../src/blocks/edit-popup/edit-popup.css);
+@import url(../src/blocks/image-popup/image-popup.css);
+@import url(../src/blocks/avatar-popup/avatar-popup.css);
+
+    /* Header */
+@import url(../src/blocks/header/header.css);
+@import url(../src/blocks/header/__logo/header__logo.css);
+
+    /* Profile */
+@import url(../src/blocks/profile/profile.css);
+@import url(../src/blocks/profile/__author-subline/profile__author-subline.css);
+@import url(../src/blocks/profile/__avatar/profile__avatar.css);
+@import url(../src/blocks/profile/__name/profile__name.css);
+@import url(../src/blocks/profile/__description/profile__description.css);
+@import url(../src/blocks/profile/__main/profile__main.css);
+@import url(../src/blocks/profile/__add-button/profile__add-button.css);
+@import url(../src/blocks/profile/__edit-button/profile__edit-button.css);
+@import url(../src/blocks/profile/__edit-avatar-button/profile__edit-avatar-button.css);
+@import url(../src/blocks/profile/__edit-avatar-button/profile__edit-avatar-button.css);
+@import url(../src/blocks/profile/__avatar-container/profile__avatar-container.css);
+
+    /* Elements */
+@import url(../src/blocks/elements/elements.css);
+@import url(../src/blocks/element/element.css);
+@import url(../src/blocks/element/__image/element__image.css);
+@import url(../src/blocks/element/__info/element__info.css);
+@import url(../src/blocks/element/__title/element__title.css);
+@import url(../src/blocks/element/__like-button/element__like-button.css);
+@import url(../src/blocks/element/__like-button/_active/element__like-button_active.css);
+@import url(../src/blocks/element/__trash-button/element__trash-button.css);
+@import url(../src/blocks/element/__like-container/element__like-container.css);
+@import url(../src/blocks/element/__like-count/element__like-count.css);
+@import url(../src/blocks/element/__trash-button/_disabled/element__trash-button_disabled.css);
+
+    /* Footer */
+@import url(../src/blocks/footer/footer.css);
+@import url(../src/blocks/footer/__copyright/footer__copyright.css);
