--- src/pages/index.css
@@ -1,68 +1,0 @@
-    /* Vendor */
-@import url(../vendor/normalize.css);
-@import url(../vendor/fonts.css);
-
-    /* Page */
-@import url(../blocks/page/page.css);
-@import url(../blocks/page/__header/page__header.css);
-@import url(../blocks/page/__profile/page__profile.css);
-@import url(../blocks/page/__elements/page__elements.css);
-@import url(../blocks/page/__footer/page__footer.css);
-
-    /* Popup */
-@import url(../blocks/popup/popup.css);
-@import url(../blocks/popup/__field/popup__field.css);
-@import url(../blocks/popup/__container/popup__container.css);
-@import url(../blocks/popup/_opened/popup_opened.css);
-@import url(../blocks/popup/__heading/popup__heading.css);
-@import url(../blocks/popup/__edit-info/popup__edit-info.css);
-@import url(../blocks/popup/__input/popup__input.css);
-@import url(../blocks/popup/__submit-button/popup__submit-button.css);
-@import url(../blocks/popup/__submit-button/_inactive/popup__submit-button_inactive.css);
-@import url(../blocks/popup/__caption/popup__caption.css);
-@import url(../blocks/popup/__figure/popup__figure.css);
-@import url(../blocks/popup/__image/popup__image.css);
-@import url(../blocks/popup/close-button/popup__close-button.css);
-@import url(../blocks/popup/__input/_type_error/popup__input_type_error.css);
-@import url(../blocks/popup/__input-error/popup__input-error.css);
-@import url(../blocks/popup/__input-error/_active/popup__input-error_active.css);
-
-
-@import url(../blocks/add-popup/add-popup.css);
-@import url(../blocks/edit-popup/edit-popup.css);
-@import url(../blocks/image-popup/image-popup.css);
-@import url(../blocks/avatar-popup/avatar-popup.css);
-
-    /* Header */
-@import url(../blocks/header/header.css);
-@import url(../blocks/header/__logo/header__logo.css);
-
-    /* Profile */
-@import url(../blocks/profile/profile.css);
-@import url(../blocks/profile/__author-subline/profile__author-subline.css);
-@import url(../blocks/profile/__avatar/profile__avatar.css);
-@import url(../blocks/profile/__name/profile__name.css);
-@import url(../blocks/profile/__description/profile__description.css);
-@import url(../blocks/profile/__main/profile__main.css);
-@import url(../blocks/profile/__add-button/profile__add-button.css);
-@import url(../blocks/profile/__edit-button/profile__edit-button.css);
-@import url(../blocks/profile/__edit-avatar-button/profile__edit-avatar-button.css);
-@import url(../blocks/profile/__edit-avatar-button/profile__edit-avatar-button.css);
-@import url(../blocks/profile/__avatar-container/profile__avatar-container.css);
-
-    /* Elements */
-@import url(../blocks/elements/elements.css);
-@import url(../blocks/element/element.css);
-@import url(../blocks/element/__image/element__image.css);
-@import url(../blocks/element/__info/element__info.css);
-@import url(../blocks/element/__title/element__title.css);
-@import url(../blocks/element/__like-button/element__like-button.css);
-@import url(../blocks/element/__like-button/_active/element__like-button_active.css);
-@import url(../blocks/element/__trash-button/element__trash-button.css);
-@import url(../blocks/element/__like-container/element__like-container.css);
-@import url(../blocks/element/__like-count/element__like-count.css);
-@import url(../blocks/element/__trash-button/_disabled/element__trash-button_disabled.css);
-
-    /* Footer */
-@import url(../blocks/footer/footer.css);
-@import url(../blocks/footer/__copyright/footer__copyright.css);
